+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.2.0
 - Added TweakScale compatibility
 - Changed from single (float) to double precision for internal computation
+- Changed sea-level isp computation to ensure it never exceeds the vacuum isp and is more strongly dependent on the thrust multiplier
 - Fixed a couple of bugs surrounding `ModuleIgnitionTankController` added mass and cost
 # 1.1.4
 - Restricted `ModuleIgnitionTankController` to only add its configured resources to its host part in the editor